+++ tests/unit/specs/src/dapps/RegisterDomainContainer/RegisterDomainContainer.spec.js
@@ -3,9 +3,7 @@ import { shallowMount } from '@vue/test-utils';
 import RegisterDomain from '@/dapps/RegisterDomain/RegisterDomain.vue';
 import BackButton from '@/layouts/InterfaceLayout/components/BackButton/BackButton.vue';
 import Vuex from 'vuex';
-import nodeList from '@/networks';
-import url from 'url';
-import Web3 from 'web3';
+import { state, getters } from '@@/helpers/mockStore';
 
 import { Tooling } from '@@/helpers';
 
@@ -33,30 +31,9 @@ describe('RegisterDomain.vue', () => {
     Vue.config.warnHandler = () => {};
     Vue.config.errorHandler = () => {};
 
-    const network = nodeList['ETH'][3];
-    const hostUrl = url.parse(network.url);
-
-    const newWeb3 = new Web3(
-      `${hostUrl.protocol}//${hostUrl.hostname}:${network.port}${
-        hostUrl.pathname
-      }`
-    );
-
-    const getters = {
-      Networks: () => {
-        return nodeList;
-      },
-      network: () => {
-        return network;
-      },
-      web3: () => {
-        return newWeb3;
-      },
-      ens: () => {}
-    };
-
     store = new Vuex.Store({
-      getters
+      getters,
+      state
     });
   });